feat: 修复表格显示
