+++ component/mini-parser/index.wxss
@@ -1,3 +1,3 @@
-.mini-parser{
+.mini-parser {
     width: 100%;
 }
