+++ component/mini-parser/render/index.json
@@ -1,6 +1,7 @@
 {
   "component": true,
   "usingComponents": {
-    "render": "./index"
+    "render": "./index",
+    "tableRender": "../table-render/index"
   }
 }
