+++ component/mini-parser/render/index.wxml
@@ -1,41 +1,35 @@
 <block wx:if="{{parserData}}">
-  <block wx:for="{{parserData}}" wx:key="id">
-    <!-- 文字 -->
-    <text wx:if="{{item.name==='text'}}" class="mini-parser-text {{item.attrs.class}}">{{item.attrs.content}}</text>
-    <!-- 图片 -->
-    <image wx:elif="{{item.name==='img'}}"
-           class="mini-parser-image {{item.attrs.class}}"
-           src="{{item.attrs.src}}"
-           style="{{item.attrs.style}}"></image>
-    <!-- 视频 -->
-    <video wx:elif="{{item.name==='video'}}"
-           class="mini-parser-video {{item.attrs.class}}"
-           src="{{item.attrs.src}}"
-           style="{{item.attrs.style}}"></video>
-    <!-- 链接 -->
-    <view wx:elif="{{item.name==='a'}}"
-          class="mini-parser-link mini-parser-{{item.display}} {{item.attrs.class}}"
-          style="{{item.attrs.style}}"
-          data-attrs="{{item.attrs}}"
-          bind:tap="handleLinkTap">
-      <render wx:if="{{item.children}}" parsedData="{{item.children}}"></render>
-    </view>
-    <!-- 表格 -->
-    <view wx:elif="{{item.name==='table'}}" class="mini-parser-table {{item.attrs.class}}">
-      <view wx:for="{{item.children}}" wx:for-item="tr" wx:key="id" class="mini-parser-tr {{tr.attrs.class}}">
-        <view wx:for="{{tr.children}}" wx:for-item="td" wx:key="id" class="mini-parser-td {{td.attrs.class}}">
-          <render wx:if="{{td.children}}" parsedData="{{td.children}}"></render>
+    <block wx:for="{{parserData}}" wx:key="id">
+        <!-- 文字 -->
+        <text wx:if="{{item.name==='text'}}" class="mini-parser-text {{item.attrs.class}}">{{item.attrs.content}}</text>
+        <!-- 图片 -->
+        <image wx:elif="{{item.name==='img'}}"
+               class="mini-parser-image {{item.attrs.class}}"
+               src="{{item.attrs.src}}"
+               style="{{item.attrs.style}}"></image>
+        <!-- 视频 -->
+        <video wx:elif="{{item.name==='video'}}"
+               class="mini-parser-video {{item.attrs.class}}"
+               src="{{item.attrs.src}}"
+               style="{{item.attrs.style}}"></video>
+        <!-- 链接 -->
+        <view wx:elif="{{item.name==='a'}}"
+              class="mini-parser-link mini-parser-{{item.display}} {{item.attrs.class}}"
+              style="{{item.attrs.style}}"
+              data-attrs="{{item.attrs}}"
+              bind:tap="handleLinkTap">
+            <render wx:if="{{item.children}}" parsedData="{{item.children}}"></render>
         </view>
-      </view>
-    </view>
-    <!-- 换行符 -->
-    <text wx:elif="{{item.name==='br'}}" class="mini-parser-br {{item.attrs.class}}">\n</text>
-    <!-- 默认 -->
-    <view wx:else
-          class="mini-parser-origin-{{item.name}} mini-parser-view mini-parser-{{item.display}} {{item.attrs.class}}"
-          style="{{item.attrs.style}}">
-      <render wx:if="{{item.children}}" parsedData="{{item.children}}"></render>
-    </view>
-  </block>
+        <!-- 表格 -->
+        <tableRender wx:elif="{{item.name==='table'}}" tableData="{{item}}"></tableRender>
+        <!-- 换行符 -->
+        <text wx:elif="{{item.name==='br'}}" class="mini-parser-br {{item.attrs.class}}">\n</text>
+        <!-- 默认 -->
+        <view wx:else
+              class="mini-parser-origin-{{item.name}} mini-parser-view mini-parser-{{item.display}} {{item.attrs.class}}"
+              style="{{item.attrs.style}}">
+            <render wx:if="{{item.children}}" parsedData="{{item.children}}"></render>
+        </view>
+    </block>
 </block>
