+++ component/mini-parser/table-render/index.json
@@ -1,0 +1,6 @@
+{
+  "component": true,
+  "usingComponents": {
+    "render": "../render/index"
+  }
+}
