+++ component/mini-parser/table-render/index.wxss
@@ -1,0 +1,20 @@
+.mini-parser-table {
+    display: table;
+    width: 100%;
+    box-sizing: border-box;
+    overflow: hidden;
+    border: 2rpx solid #dcdcdc;
+    border-width: 2rpx 0 0 2rpx;
+}
+
+.mini-parser-tr {
+    display: table-row;
+}
+
+.mini-parser-td {
+    display: table-cell;
+    text-align: start;
+    border: 2rpx solid #dcdcdc;
+    border-width: 0 2rpx 2rpx 0;
+    padding: 10rpx 15rpx;
+}
